landing/requirements.txt
confluent-kafka==2.1.1
jsonschema==4.17.3
python-dotenv==1.0.0